pgp_email: Make necessary model changes and migrations.
We add a new table UserPGP, for storing users' public keys. We also add
fields want_encrypted_emails and want_signed_emails to UserProfile,
which store user preferences for whether they want emails zulip will
send to them encrypted and/or signed. Finally, we make the corresponding
migrations and add UserPGP to the lists of tables in export.py.