pgp_email: Start implementing pgp_sign_and_encrypt functionality.
We create zerver/lib/pgp.py, with the function pgp_sign_and_encrypt
which, given a to-be-sent email message and a list of user profiles,
will encrypt and/or sign the message, if appropriate based on server
settings and user preferences, and return a list of emails ready to be
sent. We add the option to enable/disable encryption and signatures in
settings.py, by default set to False and should be kept that way until
the features are actually implemented.

At the moment trying to sign or encrypt will cause NotImplementedError
to be raised, but works (basically doing nothing except setting the To
header) in the most basic case where we don't want any
encryption/signatures.