pgp_email: Begin implementing pgp, by overriding .message().
To have the email correctly formatted according to PGP/MIME (RFC3156),
we need full control over the various MIME parts and headers, that we
don't have through EmailMultiAlternatives. When sending an email,
Django invokes the .message() method of EmailMessage which converts it
and returns an object of type used by the python "email" library.
This is the stage at which we have the desired level of control.
(See django.core.mail.message for details.)
Thus, encryption and signing need to happen by overriding .message().