pgp_email: Implement PGP/MIME formatting of emails.
When .send() is invoked for an email message, Django sends the output of
.message() through the apprioprate backend. This .message() output is
essentially a MIME object from the python "email" module. If we're
sending an encrypted/signed email, this object needs to have an
appropriate PGP/MIME structure, as described in RFC3156. In this commit,
we implement the creation of this structure. Instead of actual
ciphertexts/signatures, we insert dummy strings for now, as this last
part of the process isn't implemented yet.

Appropriate tests are added to test_pgp.py.