rate_limiter: Change structure of reset_times in the tornado backend.
This can significantly improve performance in the future if this code
is used in scenarios where it'll be common for various keys to have
different rules applied to them.
With this change, when garbage collecting for rule X we'll only iterate
over keys that had rule X applied to them - instead of iterating over
all keys unconditionally.