tests: Add tests for "Sign in with Apple".
"Sign in with Apple" being a bit different from other backends
a different `social_auth_test` is added to it.

Three new tests were also added.
* `test_get_apple_locale` to test the locale sent
to get the "Sign in with Apple" button in a particular language
is correct.
* `test_auth_registration_with_no_name_sent_from_apple` to test
the case where apple doesn't send "name" of user if the user
tries to register after a failed first attempt.
* `test_validate_state` to test the behaviour of overriden PSA's
`validate_state` function.