presence: Rewrite the backend data model.
This implements the core of the rewrite described in:

For the backend data model for UserPresence to one that supports much
more efficient queries and is more correct around handling of multiple
clients.  The main loss of functionality is that we no longer track
which Client sent presence data (so we will no longer be able to say
using UserPresence "the user was last online on their desktop 15
minutes ago, but was online with their phone 3 minutes ago").  If we
consider that information important for the occasional investigation
query, we have can construct that answer data via UserActivity
already.  It's not worth making Presence much more expensive/complex
to support it.

For slim_presence clients, this sends the same data format we sent
before, albeit with less complexity involved in constructing it.  Note
that we at present will always send both last_active_time and
last_connected_time; we may revisit that in the future.

TODO:
* Needs the `notify_presence_changed` event to be fixed.
* Unit tests are broken; some need to be changed to check
!slim_presence behavior, and others make the assumption that user
accounts having their first UserPresence row added don't update
`last_active_time` is STATUS=idle; we need to think this through.
* Thoughtful decisions around various timedelta windows.