auth: Prevent unauthorized signup in ldap + external auth cases.
If:
- ZulipLDAPAuthBackend and an external authentication backend (any aside
  of ZulipLDAPAuthBackend and EmailAuthBackend) are the only ones
  enabled in AUTHENTICATION_BACKENDS in /etc/zulip/settings.py
- The organization permissions don't require invitations to join

An attacker can create a new account in the organization with an
arbitrary email address in their control that's not in the
organization's LDAP directory.

The impact is limited to installations which have this specific
combination of authentication backends as described above in addition to
having Invitations are required for joining this organization
organization permission disabled.

Bug introduced in 74dd21c8fa132f0e55691fc05cba812348a11d55